refactor(core): promote pkg/proxy to internal/core/proxy (no behavior change)
