+++ cmd/databricks-claude/serve_session.go
@@ -14,9 +14,9 @@ import (
 	"github.com/IceRhymers/databricks-agents/internal/core/authcheck"
 	"github.com/IceRhymers/databricks-agents/internal/core/lifecycle"
 	"github.com/IceRhymers/databricks-agents/internal/core/portbind"
+	"github.com/IceRhymers/databricks-agents/internal/core/proxy"
 	"github.com/IceRhymers/databricks-agents/internal/core/refcount"
 	"github.com/IceRhymers/databricks-agents/pkg/health"
-	"github.com/IceRhymers/databricks-agents/pkg/proxy"
 )
 
 // runServeSession implements `databricks-claude serve --session-mode`. The
