+++ internal/core/proxy/anthropic/types.go
