+++ internal/core/proxy/checks_test.go
