+++ internal/core/proxy/checks_unix.go
