+++ internal/core/proxy/checks_windows.go
