+++ internal/core/proxy/proxy_test.go
