+++ internal/core/proxy/responses_rewriter.go
