+++ internal/core/proxy/responses_rewriter_test.go
