+++ internal/core/proxy/sanitize.go
