+++ internal/core/proxy/sanitize_test.go
