+++ internal/core/proxy/sse_rewriter.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/IceRhymers/databricks-agents/pkg/proxy/anthropic"
+	"github.com/IceRhymers/databricks-agents/internal/core/proxy/anthropic"
 	"github.com/IceRhymers/databricks-agents/pkg/websearch"
 )
