+++ internal/core/proxy/sse_rewriter_test.go
