+++ internal/core/proxy/websearch_handler.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/IceRhymers/databricks-agents/pkg/proxy/anthropic"
+	"github.com/IceRhymers/databricks-agents/internal/core/proxy/anthropic"
 	"github.com/IceRhymers/databricks-agents/pkg/websearch"
 )
